refactor from migration folder
