--- deepretro/migration/__init__.py
@@ -1,5 +1,0 @@
-"""Holding area for src → deepretro ports before they are merged into the main package.
-
-See NOTES.txt for the rule: do not duplicate modules that already exist under
-deepretro — import those; only keep gaps and glue here.
-"""
