--- deepretro/migration/job_context.py
@@ -1,4 +1,0 @@
-import contextvars
-
-# Create a context variable for the job ID
-logger: contextvars.ContextVar = contextvars.ContextVar("logger")
