--- deepretro/migration/rec_prithvi.py
@@ -1,3 +1,0 @@
-"""Shim — rec_run_prithvi moved to deepretro.algorithms.autosolve.rec_run."""
-
-from deepretro.algorithms.autosolve import rec_run as rec_run_prithvi  # noqa: F401
